+++ components/radio/src/composables/radio-custom.ts
@@ -44,6 +44,8 @@ export const useRadioCustomStyle = (props: RadioProps) => {
         cls.push('tn-border')
         if (selected && borderColorClass.value) cls.push(borderColorClass.value)
         else cls.push('tn-gray-disabled_border')
+      } else {
+        cls.push(ns.m('no-border'))
       }
 
       if (props.customClass) cls.push(props.customClass)